+++ src/components/shared/ReactFlow/FlowGraph/utils/replaceTaskArgumentsInGraphSpec.test.ts
@@ -1,6 +1,7 @@
 import { describe, expect, it } from "vitest";
 
-import type { ArgumentType, GraphSpec } from "./componentSpec";
+import type { ArgumentType, GraphSpec } from "@/utils/componentSpec";
+
 import replaceTaskArgumentsInGraphSpec from "./replaceTaskArgumentsInGraphSpec";
 
 describe("replaceTaskArgumentsInGraphSpec", () => {
