+++ src/components/shared/Status/index.ts
@@ -1,4 +1,3 @@
 export { default as StatusIcon } from "./StatusIcon";
 export { default as StatusText } from "./StatusText";
 export { default as StatusBar } from "./TaskStatusBar";
-export * from "./utils";
